test_get_indices_of_matching_sort_exprs_with_order_eq
